repository creation in filesystem and database

+++ src/main/java/com/suhailkandanur/fresco/service/impl/RepositoryServiceImpl.java
@@ -1,6 +1,5 @@
 package com.suhailkandanur.fresco.service.impl;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.suhailkandanur.fresco.configuration.FrescoConfiguration;
 import com.suhailkandanur.fresco.dataaccess.RepositoryDAO;
 import com.suhailkandanur.fresco.entity.Repository;
@@ -10,7 +9,6 @@ import com.suhailkandanur.fresco.util.JsonUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,6 +19,7 @@ import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
+import java.util.Optional;
 
 /**
  * Created by suhail on 2016-12-02.
@@ -41,6 +40,16 @@ public class RepositoryServiceImpl implements RepositoryService {
     public void processMessage(String message) throws IOException {
         Map<String, String> request = JsonUtils.convertStrToJson(message, HashMap.class);
         logger.info("received message: {}, object: {}", message, request);
+        Optional<Repository> repositoryInDb = repositoryDAO.findRepositoryByName(request.get("name"));
+        if (repositoryInDb.isPresent()) {
+            logger.error("repository with name '{}' already present, cannot create a new one", request.get("name"));
+            return;
+        }
+        Repository repository = new Repository(request.get("name"),
+                request.get("description"),
+                Long.valueOf(request.getOrDefault("quota", "1000000")));
+        createRepository(repository);
+        writeEntryToDatabase(repository);
     }
 
     void createRepository(Repository repository) throws IOException {
@@ -61,9 +70,15 @@ public class RepositoryServiceImpl implements RepositoryService {
         Path storesRootPath = repositoryRoot.resolve("stores");
         Files.createDirectories(storesRootPath);
 
-        Path metaDataFileTmp = repositoryRoot.resolve(".meta.inf.tmp");
-        if (FileUtils.writeToFile(metaDataFileTmp, JsonUtils.convertObjectToJsonStr(repository)) != null) {
-            Files.move(metaDataFileTmp, repositoryRoot.resolve("meta.inf"));
+        Path metaDataFile = repositoryRoot.resolve("meta.inf");
+        if (Files.notExists(metaDataFile)) {
+            Path metaDataFileTmp = repositoryRoot.resolve(".meta.inf.tmp");
+            if(Files.exists(metaDataFileTmp)) Files.delete(metaDataFileTmp);
+            if (FileUtils.writeToFile(metaDataFileTmp, JsonUtils.convertObjectToJsonStr(repository)) != null) {
+                Files.move(metaDataFileTmp, metaDataFile);
+                if (Files.exists(metaDataFileTmp)) Files.delete(metaDataFileTmp);
+                logger.info("successfully created meta data file {}", metaDataFile);
+            }
         }
     }
